Stripped quotations from names and states
The CSV is already being read as a large string, and once the values / columns
are split, they are treated as strings. The quotations being around the values
would lead to them being escaped to maintain the quotations.

The escaped string sequences would also spill over into the hash-maps, which
greatly differ from the original intention.

Test failing.